refactor(holochain_zome_types): dissolve dht_v2 into action/op/record
