+++ crates/holochain_zome_types/src/metadata.rs
@@ -1,5 +1,5 @@
 //! Metadata types for use in wasm
-use crate::dht_v2::Record;
+use crate::record::Record;
 use crate::record::SignedActionHashed;
 use crate::validate::ValidationStatus;
 use crate::Entry;
@@ -26,9 +26,9 @@ pub struct RecordDetails {
     pub record: Record,
     /// The validation status of this record.
     pub validation_status: ValidationStatus,
-    /// Any [`Delete`](crate::dht_v2::ActionData::Delete) on this record.
+    /// Any [`Delete`](crate::action::ActionData::Delete) on this record.
     pub deletes: Vec<SignedActionHashed>,
-    /// Any [`Update`](crate::dht_v2::ActionData::Update) on this record.
+    /// Any [`Update`](crate::action::ActionData::Update) on this record.
     pub updates: Vec<SignedActionHashed>,
 }
 
@@ -39,8 +39,8 @@ pub struct EntryDetails {
     pub entry: Entry,
     /// ## Create relationships.
     /// These are the actions that created this entry.
-    /// They can be either a [`Create`](crate::dht_v2::ActionData::Create) or an
-    /// [`Update`](crate::dht_v2::ActionData::Update) action
+    /// They can be either a [`Create`](crate::action::ActionData::Create) or an
+    /// [`Update`](crate::action::ActionData::Update) action
     /// where the `entry_hash` field is the hash of
     /// the above entry.
     ///
