+++ crates/holochain_zome_types/src/prelude.rs
@@ -27,7 +27,6 @@ pub use crate::op::*;
 pub use crate::properties::*;
 pub use crate::query::ChainQueryFilter as QueryFilter;
 pub use crate::query::*;
-pub use crate::record::*;
 pub use crate::request::*;
 pub use crate::schedule::*;
 pub use crate::signal::*;
@@ -44,11 +43,26 @@ pub use crate::zome_io::ExternIO;
 pub use crate::zome_io::*;
 pub use holochain_integrity_types::prelude::*;
 
+// `capability::CapAccess` (data-carrying grant access) and `action::CapAccess`
+// (the `CapGrant.cap_access` column discriminant) share a name; an explicit
+// re-export takes priority over the glob re-exports above, resolving the
+// ambiguity in favor of the grant-access type at the crate root.
+// `action::CapAccess` remains reachable via its `action::` path.
+pub use crate::capability::CapAccess;
+
+// `crate::action::SignedActionHashed` and
+// `holochain_integrity_types::action::SignedActionHashed` are the same
+// underlying alias (`SignedHashed<Action>`) defined in two crates; an
+// explicit re-export takes priority over the glob re-exports above,
+// resolving the ambiguity in favor of this crate's own alias, which is the
+// path downstream consumers use.
+pub use crate::action::SignedActionHashed;
+
 // Bring the validation `Op` and its variant structs into the prelude so
 // validators and inline zomes decode the `Op` the host encodes, and
 // `MustGetAgentActivityResponse` carries the `RegisterAgentActivity` payload
 // directly.
-pub use crate::dht_v2::{
+pub use crate::op::{
     Op, RegisterAgentActivity, RegisterCreateLink, RegisterDelete, RegisterDeleteLink,
     RegisterUpdate, StoreEntry, StoreRecord,
 };
