+++ crates/holochain_zome_types/src/record.rs
@@ -3,9 +3,7 @@
 // Record-support types (`RecordEntry`, `SignedHashed`, etc.).
 pub use holochain_integrity_types::record::*;
 
-/// The canonical chain record: a signed, hashed action plus its entry.
-pub use crate::dht_v2::Record;
 /// An action with its signature (no hash).
-pub use crate::dht_v2::SignedAction;
+pub use crate::action::SignedAction;
 /// An action that is both content-addressed and signed.
-pub use crate::dht_v2::SignedActionHashed;
+pub use crate::action::SignedActionHashed;
